@@ -1,11 +1,19 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <PackageId>Weyhd.AspNetCore.Authentication.WXWork</PackageId>
+    <Authors>Weyhd</Authors>
+    <Product>Weyhd.AspNetCore.Authentication.WXWork</Product>
+    <Description>适用于ASP.NET Core3.1 的 OAuth2 认证中间件，使Web应用程序支持用WXWork快捷登录。</Description>
+    <RepositoryType>github</RepositoryType>
+    <PackageTags>WXWork, security, aspnet core, authentication</PackageTags>
+    <Version>3.1.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.OAuth" Version="2.0.0" />
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="JetBrains.Annotations" Version="2019.1.3" PrivateAssets="All" />
   </ItemGroup>
 
 </Project>
