@@ -1,47 +1,10 @@
-﻿using Microsoft.AspNetCore.Authentication;
+﻿using JetBrains.Annotations;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Weixin;
 using System;
 
 
 namespace Microsoft.Extensions.DependencyInjection
-{
-    /// <summary> 
-    /// </summary>
-    public static class QQAuthenticationExtensions
-    {
-        /// <summary> 
-        /// </summary>
-        public static AuthenticationBuilder AddWeixinAuthentication(this AuthenticationBuilder builder)
-        {
-            return builder.AddWeixinAuthentication(WeixinAuthenticationDefaults.AuthenticationScheme, WeixinAuthenticationDefaults.DisplayName, options => { });
-        }
-
-        /// <summary> 
-        /// </summary>
-        public static AuthenticationBuilder AddWeixinAuthentication(this AuthenticationBuilder builder, Action<WeixinAuthenticationOptions> configureOptions)
-        {
-            return builder.AddWeixinAuthentication(WeixinAuthenticationDefaults.AuthenticationScheme, WeixinAuthenticationDefaults.DisplayName, configureOptions);
-        }
-
-        /// <summary> 
-        /// </summary>
-        public static AuthenticationBuilder AddWeixinAuthentication(this AuthenticationBuilder builder, string authenticationScheme, Action<WeixinAuthenticationOptions> configureOptions)
-        {
-            return builder.AddWeixinAuthentication(authenticationScheme, WeixinAuthenticationDefaults.DisplayName, configureOptions);
-        }
-
-        /// <summary> 
-        /// </summary>
-        public static AuthenticationBuilder AddWeixinAuthentication(this AuthenticationBuilder builder, string authenticationScheme, string displayName, Action<WeixinAuthenticationOptions> configureOptions)
-        {
-            return builder.AddOAuth<WeixinAuthenticationOptions, WeixinAuthenticationHandler>(authenticationScheme, displayName, configureOptions);
-        }
-    }
-}
-
-#region Old Code 
-
-namespace Microsoft.AspNetCore.Builder
 {
     /// <summary>
     /// Extension methods to add Weixin authentication capabilities to an HTTP application pipeline.
@@ -49,31 +12,60 @@ namespace Microsoft.AspNetCore.Builder
     public static class WeixinAuthenticationExtensions
     {
         /// <summary>
-        /// Adds the <see cref="WeixinAuthenticationMiddleware"/> middleware to the specified
-        /// <see cref="IApplicationBuilder"/>, which enables Weixin authentication capabilities.
+        /// Adds <see cref="WeixinAuthenticationHandler"/> to the specified
+        /// <see cref="AuthenticationBuilder"/>, which enables Weixin authentication capabilities.
         /// </summary>
-        /// <param name="app">The <see cref="IApplicationBuilder"/> to add the middleware to.</param>
-        /// <param name="options">A <see cref="WeixinAuthenticationOptions"/> that specifies options for the middleware.</param>
+        /// <param name="builder">The authentication builder.</param>
         /// <returns>A reference to this instance after the operation has completed.</returns>
-        public static IApplicationBuilder UseWeixinAuthentication(this IApplicationBuilder app, WeixinAuthenticationOptions options)
+        public static AuthenticationBuilder AddWeixin([NotNull] this AuthenticationBuilder builder)
         {
-            throw new NotSupportedException("This method is no longer supported, see https://go.microsoft.com/fwlink/?linkid=845470");
+            return builder.AddWeixin(WeixinAuthenticationDefaults.AuthenticationScheme, options => { });
         }
 
         /// <summary>
-        /// Adds the <see cref="WeixinAuthenticationMiddleware"/> middleware to the specified
-        /// <see cref="IApplicationBuilder"/>, which enables Weixin authentication capabilities.
+        /// Adds <see cref="WeixinAuthenticationHandler"/> to the specified
+        /// <see cref="AuthenticationBuilder"/>, which enables Weixin authentication capabilities.
         /// </summary>
-        /// <param name="app">The <see cref="IApplicationBuilder"/> to add the middleware to.</param>
-        /// <param name="configuration">An action delegate to configure the provided <see cref="WeixinAuthenticationOptions"/>.</param>
+        /// <param name="builder">The authentication builder.</param>
+        /// <param name="configuration">The delegate used to configure the OpenID 2.0 options.</param>
         /// <returns>A reference to this instance after the operation has completed.</returns>
-        public static IApplicationBuilder UseWeixinAuthentication(this IApplicationBuilder app, Action<WeixinAuthenticationOptions> configuration)
+        public static AuthenticationBuilder AddWeixin(
+            [NotNull] this AuthenticationBuilder builder,
+            [NotNull] Action<WeixinAuthenticationOptions> configuration)
         {
-            throw new NotSupportedException("This method is no longer supported, see https://go.microsoft.com/fwlink/?linkid=845470");
+            return builder.AddWeixin(WeixinAuthenticationDefaults.AuthenticationScheme, configuration);
+        }
+
+        /// <summary>
+        /// Adds <see cref="WeixinAuthenticationHandler"/> to the specified
+        /// <see cref="AuthenticationBuilder"/>, which enables Weixin authentication capabilities.
+        /// </summary>
+        /// <param name="builder">The authentication builder.</param>
+        /// <param name="scheme">The authentication scheme associated with this instance.</param>
+        /// <param name="configuration">The delegate used to configure the Weixin options.</param>
+        /// <returns>The <see cref="AuthenticationBuilder"/>.</returns>
+        public static AuthenticationBuilder AddWeixin(
+            [NotNull] this AuthenticationBuilder builder, [NotNull] string scheme,
+            [NotNull] Action<WeixinAuthenticationOptions> configuration)
+        {
+            return builder.AddWeixin(scheme, WeixinAuthenticationDefaults.DisplayName, configuration);
+        }
+
+        /// <summary>
+        /// Adds <see cref="WeixinAuthenticationHandler"/> to the specified
+        /// <see cref="AuthenticationBuilder"/>, which enables Weixin authentication capabilities.
+        /// </summary>
+        /// <param name="builder">The authentication builder.</param>
+        /// <param name="scheme">The authentication scheme associated with this instance.</param>
+        /// <param name="caption">The optional display name associated with this instance.</param>
+        /// <param name="configuration">The delegate used to configure the Weixin options.</param>
+        /// <returns>The <see cref="AuthenticationBuilder"/>.</returns>
+        public static AuthenticationBuilder AddWeixin(
+            [NotNull] this AuthenticationBuilder builder,
+            [NotNull] string scheme, [CanBeNull] string caption,
+            [NotNull] Action<WeixinAuthenticationOptions> configuration)
+        {
+            return builder.AddOAuth<WeixinAuthenticationOptions, WeixinAuthenticationHandler>(scheme, caption, configuration);
         }
     }
-
-
 }
-
-#endregion
